+++ src/models/AccrodionPriceModel.tsx
@@ -1,0 +1,4 @@
+export interface AccrodionPriceModel {
+  item: string;
+  price: number;
+}